+++ src/model/context.rs
@@ -1,11 +1,14 @@
 //! All objects related to context
+use super::device::Device;
+use super::PlayingItem;
+use crate::model::{
+    from_millisecond_timestamp, from_option_duration_ms, to_millisecond_timestamp,
+    to_option_duration_ms, CurrentlyPlayingType, DisallowKey, RepeatState, Type,
+};
+use chrono::{DateTime, Utc};
 use serde::{Deserialize, Deserializer, Serialize};
 use std::collections::HashMap;
-
-use super::device::Device;
-use super::track::FullTrack;
-use super::PlayingItem;
-use crate::model::{CurrentlyPlayingType, DisallowKey, RepeatState, Type};
+use std::time::Duration;
 /// Context object
 ///
 /// [Reference](https://developer.spotify.com/web-api/get-the-users-currently-playing-track/)
@@ -18,26 +21,24 @@ pub struct Context {
     pub _type: Type,
 }
 
-/// Simplified playing context
-///
-/// [Reference](https://developer.spotify.com/web-api/get-the-users-currently-playing-track/)
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-pub struct SimplifiedPlayingContext {
-    pub context: Option<Context>,
-    pub timestamp: u64,
-    pub progress_ms: Option<u32>,
-    pub is_playing: bool,
-    pub item: Option<FullTrack>,
-}
-
 /// Currently playing object
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/player/get-the-users-currently-playing-track/)
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct CurrentlyPlayingContext {
     pub context: Option<Context>,
-    pub timestamp: u64,
-    pub progress_ms: Option<u32>,
+    #[serde(
+        deserialize_with = "from_millisecond_timestamp",
+        serialize_with = "to_millisecond_timestamp"
+    )]
+    pub timestamp: DateTime<Utc>,
+    #[serde(default)]
+    #[serde(
+        deserialize_with = "from_option_duration_ms",
+        serialize_with = "to_option_duration_ms",
+        rename = "progress_ms"
+    )]
+    pub progress: Option<Duration>,
     pub is_playing: bool,
     pub item: Option<PlayingItem>,
     pub currently_playing_type: CurrentlyPlayingType,
@@ -50,8 +51,18 @@ pub struct CurrentPlaybackContext {
     pub repeat_state: RepeatState,
     pub shuffle_state: bool,
     pub context: Option<Context>,
-    pub timestamp: u64,
-    pub progress_ms: Option<u32>,
+    #[serde(
+        deserialize_with = "from_millisecond_timestamp",
+        serialize_with = "to_millisecond_timestamp"
+    )]
+    pub timestamp: DateTime<Utc>,
+    #[serde(default)]
+    #[serde(
+        deserialize_with = "from_option_duration_ms",
+        serialize_with = "to_option_duration_ms",
+        rename = "progress_ms"
+    )]
+    pub progress: Option<Duration>,
     pub is_playing: bool,
     pub item: Option<PlayingItem>,
     pub currently_playing_type: CurrentlyPlayingType,